Preconditioners

ExtendableSparse.AbstractExtendablePreconditioner — Type

Abstract type for an extendabe preconditoioner working together with ExtandableSparseMatrix. Still beta, so not in the published documentation.

Any such preconditioner should have the following fields

  extmatrix
  pattern_timestamp

and methods

  update!(precon)

The idea is that, depending if the matrix pattern has changed, different steps are needed to update the preconditioner.

Moreover, they have the ExtendableSparseMatrix as a field, ensuring consistency after construction.

ExtendableSparse.ILU0Preconditioner — Type

mutable struct ILU0Preconditioner{Tv, Ti} <: ExtendableSparse.AbstractExtendablePreconditioner{Tv,Ti}

ILU(0) Preconditioner

ExtendableSparse.ILU0Preconditioner — Method

Constructor for ILU(0) preconditioner

ExtendableSparse.ILU0Preconditioner — Method

Constructor for ILU(0) preconditioner

ExtendableSparse.update! — Method

Update ILU(0) preconditioner

LinearAlgebra.ldiv! — Method

ldiv!(precon, v)

Inplace solve of preconditioning system for ILU(0)

LinearAlgebra.ldiv! — Method

ldiv!(u, precon, v)

Solve preconditioning system for ILU(0)

ExtendableSparse.JacobiPreconditioner — Type

struct JacobiPreconditioner{Tv, Ti} <: ExtendableSparse.AbstractExtendablePreconditioner{Tv,Ti}

Jacobi preconditoner

ExtendableSparse.JacobiPreconditioner — Method

Construct Jacobi preconditoner

ExtendableSparse.JacobiPreconditioner — Method

Construct Jacobi preconditoner

ExtendableSparse.ParallelJacobiPreconditioner — Type

struct ParallelJacobiPreconditioner{Tv, Ti} <: ExtendableSparse.AbstractExtendablePreconditioner{Tv,Ti}

Parallel Jacobi preconditioner

ExtendableSparse.ParallelJacobiPreconditioner — Method

Parallel Jacobi preconditioner construtor

ExtendableSparse.ParallelJacobiPreconditioner — Method

Parallel Jacobi preconditioner construtor

ExtendableSparse.update! — Method

update!(precon)

Update Jacobi preconditoner

ExtendableSparse.update! — Method

update!(precon)

Parallel Jacobi preconditioner update

LinearAlgebra.ldiv! — Method

ldiv!(u, precon, v)

Solve Jacobi preconditioning system

LinearAlgebra.ldiv! — Method

ldiv!(u, precon, v)

Parallel Jacobi preconditioner solve

LinearAlgebra.ldiv! — Method

ldiv!(precon, v)

Inplace solve Jacobi preconditioning system

LinearAlgebra.ldiv! — Method

ldiv!(precon, v)

Parallel Jacobi preconditioner inplace solve

ExtendableSparse.simple! — Method

simple!(u,A,b;
                 abstol::Real = zero(real(eltype(b))),
                 reltol::Real = sqrt(eps(real(eltype(b)))),
                 log=false,
                 maxiter=100,
                 P=nothing
                 ) -> solution, [history]

Simple iteration scheme $u_{i+1}= u_i - P^{-1} (A u_i -b)$ with similar API as the methods in IterativeSolvers.jl.