API

API

Types and Constants

ExtendableSparse.SparseMatrixExtension — Type.

mutable struct SparseMatrixExtension{Tv, Ti<:Integer} <: SparseArrays.AbstractSparseArray{Tv,Ti<:Integer,2}

Struct to hold extension data of sparse matrix.

m::Integer

Number of rows

n::Integer

Number of columns

nnz::Integer

Number of nonzeros

colptr::Array{Ti,1} where Ti<:Integer

Linked list of column entries. Initial length is n, it grows with each new entry.

colptr[index] contains the next index in the list or zero, terminating the list which starts at index 1<=j<=n for each column j.

rowval::Array{Ti,1} where Ti<:Integer

Row numbers. For each index it contains the zero (initial state) or the row numbers corresponding to the column entry list in colptr.

Initial length is n, it grows with each new entry.

nzval::Array{Tv,1} where Tv

Nonzero entry values correspondin to each pair (colptr[index],rowval[index])

Initial length is n, it grows with each new entry.

source

ExtendableSparse.SparseMatrixExtension — Method.

Constructor of empty extension

source

ExtendableSparse.ExtendableSparseMatrix — Type.

mutable struct ExtendableSparseMatrix{Tv, Ti<:Integer} <: SparseArrays.AbstractSparseArray{Tv,Ti<:Integer,2}

Extendable sparse matrix. A nonzero entry of this matrix is contained either in cscmatrix, or in extmatrix, never in both.

cscmatrix::SparseArrays.SparseMatrixCSC

Final matrix data

extmatrix::Union{Nothing, SparseMatrixExtension{Tv,Ti}} where Ti<:Integer where Tv

Intermediate structure holding data of extension

source

ExtendableSparse.ExtendableSparseMatrix — Method.

ExtendableSparseMatrix(m::Integer, n::Integer) -> ExtendableSparseMatrix{Float64,Int64}

Create empty ExtendablSparseMatrix. This is a pendant to spzeros.

source

ExtendableSparse.ExtendableSparseMatrix — Method.

ExtendableSparseMatrix(M)

Create ExtendablSparseMatrix from sparse matrix

source

ExtendableSparse.ExtendableSparseMatrix — Method.

ExtendableSparseMatrix(m, n)

Create empty ExtendablSparseMatrix.

source

ExtendableSparse.ExtendableSparseMatrix — Method.

ExtendableSparseMatrix(?, ?, m, n)

Create empty ExtendablSparseMatrix.

source

ExtendableSparse.ExtendableSparseMatrix — Method.

ExtendableSparseMatrix(?::Tv, m::Tv, n::Tv) -> ExtendableSparseMatrix{_1,_2} where _2 where _1

Create empty ExtendablSparseMatrix. This is a pendant to spzeros.

source

Methods

ExtendableSparse.flush! — Method.

Dummy flush! method for Sparse matrix extension. Just used in thest methods

source

SparseArrays.nnz — Method.

nnz(E::SparseMatrixExtension) -> Integer

Return number of nonzero entries.

source

ExtendableSparse.colptrs — Method.

colptrs(E::ExtendableSparseMatrix) -> Array{Ti,1} where Ti<:Integer

Flush and delegate to cscmatrix.

source

ExtendableSparse.flush! — Method.

If there are new entries in extension, create new CSC matrix and reset extension.

source

ExtendableSparse.xcolptrs — Method.

xcolptrs(E::ExtendableSparseMatrix) -> Array{Ti,1} where Ti<:Integer

Flush and delegate to cscmatrix.

source

SparseArrays.nnz — Method.

nnz(E::ExtendableSparseMatrix) -> Any

Number of nonzeros of ExtendableSparseMatrix.

source

ExtendableSparse.sprand! — Method.

sprand!(A, xnnz)

Fill empty sparse matrix A with random nonzero elements from interval [1,2] using incremental assembly.

source

Base.getindex — Method.

getindex(E, i, j)

Return value stored for entry or zero if not found

source

Base.setindex! — Method.

setindex!(E, _v, _i, _j)

Update value of existing entry, otherwise extend matrix.

source

Base.size — Method.

size(E::SparseMatrixExtension) -> Tuple{Integer,Integer}

Return tuple containing size of the matrix.

source

Base.getindex — Method.

getindex(M, i, j)

Find index in CSC matrix and return value, if it exists. Otherwise, return value from extension.

source

Base.setindex! — Method.

setindex!(M, v, i, j)

Find index in CSC matrix and set value if it exists. Otherwise, set index in extension.

source

Base.size — Method.

size(E::ExtendableSparseMatrix) -> Tuple{Int64,Int64}

Size of ExtendableSparseMatrix.

source

ExtendableSparse.findindex — Method.

findindex(S::T, i::T, j::T)

Return index corresponding to entry (i,j) in the array of nonzeros, if the entry exists, otherwise, return 0.

source

LinearAlgebra.ldiv! — Method.

ldiv!(r::AbstractArray{T,1} where T, E::ExtendableSparseMatrix, x::AbstractArray{T,1} where T)

Delegating Matrix ldiv

source

LinearAlgebra.ldiv! — Method.

ldiv!(r::AbstractArray{T,2} where T, E::ExtendableSparseMatrix, x::AbstractArray{T,2} where T)

Delegating Matrix ldiv

source

LinearAlgebra.lu — Method.

lu(E::ExtendableSparseMatrix) -> Any

Delegating LU factorization.

source

LinearAlgebra.mul! — Method.

mul!(r::AbstractArray{T,1} where T, E::ExtendableSparseMatrix, x::AbstractArray{T,1} where T) -> Any

Delegating Matrix multiplication

source

LinearAlgebra.mul! — Method.

mul!(r::AbstractArray{T,2} where T, E::ExtendableSparseMatrix, x::AbstractArray{T,2} where T) -> Any

Delegating Matrix multiplication

source

SparseArrays.findnz — Method.

findnz(E::ExtendableSparseMatrix) -> Tuple{Any,Any,Any}

Flush and delegate to cscmatrix.

source

SparseArrays.nonzeros — Method.

nonzeros(E::ExtendableSparseMatrix) -> Array{Tv,1} where Tv

Flush and delegate to cscmatrix.

source

SparseArrays.rowvals — Method.

rowvals(E::ExtendableSparseMatrix) -> Array{Ti,1} where Ti<:Integer

Flush and delegate to cscmatrix.

source