API

API

Types and Constants

ExtendableSparse.SparseMatrixLNK — Type.

mutable struct SparseMatrixLNK{Tv, Ti<:Integer} <: SparseArrays.AbstractSparseArray{Tv,Ti<:Integer,2}

Struct to hold sparse matrix in the linked list format.

Modeled after the linked list sparse matrix format described in the whitepaper and the SPARSEKIT2 source code by Y. Saad. He writes "This is one of the oldest data structures used for sparse matrix computations."

The relevant source formats.f is also available in the debian/science gitlab.

Probably this format was around already in SPARSPAK by E.Chu, A.George and J.Liu, however this is hard to verify, as it indeed appears that the source code of SPARSPAK vanished from the internet.

The advantage of the linked list structure is the fact that upon insertion of a new entry, the arrays describing the structure grow at their respective ends and can be conveniently updated via push!. No copying of existing data is necessary.

m::Integer

Number of rows

n::Integer

Number of columns

nnz::Integer

Number of nonzeros

nentries::Integer

Length of arrays

colptr::Array{Ti,1} where Ti<:Integer

Linked list of column entries. Initial length is n, it grows with each new entry.

colptr[index] contains the next index in the list or zero, in the later case terminating the list which starts at index 1<=j<=n for each column j.

rowval::Array{Ti,1} where Ti<:Integer

Row numbers. For each index it contains the zero (initial state) or the row numbers corresponding to the column entry list in colptr.

Initial length is n, it grows with each new entry.

nzval::Array{Tv,1} where Tv

Nonzero entry values correspondin to each pair (colptr[index],rowval[index])

Initial length is n, it grows with each new entry.

source

ExtendableSparse.SparseMatrixLNK — Method.

SparseMatrixLNK(csc)

Constructor from SparseMatrixCSC.

source

ExtendableSparse.SparseMatrixLNK — Method.

SparseMatrixLNK(m, n)

Constructor of empty matrix.

source

ExtendableSparse.SparseMatrixLNK — Method.

SparseMatrixLNK(valuetype, indextype, m, n)

Constructor of empty matrix.

source

ExtendableSparse.SparseMatrixLNK — Method.

SparseMatrixLNK(m, n)

Constructor of empty matrix.

source

ExtendableSparse.SparseMatrixLNK — Method.

SparseMatrixLNK(valuetype, m, n)

Constructor of empty matrix.

source

ExtendableSparse.ExtendableSparseMatrix — Type.

mutable struct ExtendableSparseMatrix{Tv, Ti<:Integer} <: SparseArrays.AbstractSparseArray{Tv,Ti<:Integer,2}

Extendable sparse matrix. A nonzero entry of this matrix is contained either in cscmatrix, or in lnkmatrix, never in both.

cscmatrix::SparseArrays.SparseMatrixCSC

Final matrix data

lnkmatrix::Union{Nothing, SparseMatrixLNK{Tv,Ti}} where Ti<:Integer where Tv

Linked list structure holding data of extension

pattern_timestamp::Float64

Time stamp of last pattern update

source

ExtendableSparse.ExtendableSparseMatrix — Method.

ExtendableSparseMatrix(m, n)

Create empty ExtendableSparseMatrix. This is a pendant to spzeros.

source

ExtendableSparse.ExtendableSparseMatrix — Method.

Create ExtendableSparseMatrix from SparseMatrixCSC

source

ExtendableSparse.ExtendableSparseMatrix — Method.

ExtendableSparseMatrix(m, n)

Create empty ExtendableSparseMatrix.

source

ExtendableSparse.ExtendableSparseMatrix — Method.

Create empty ExtendableSparseMatrix.

source

ExtendableSparse.ExtendableSparseMatrix — Method.

Create empty ExtendablSparseMatrix. This is a pendant to spzeros.

source

SparseArrays.SparseMatrixCSC — Method.

Constructor from SparseMatrixCSC.

source

Methods

ExtendableSparse.flush! — Method.

Dummy flush! method for SparseMatrixLNK. Just used in test methods

source

ExtendableSparse.updateindex! — Method.

updateindex!(lnk, op, _v, _i, _j)

Update element of the matrix with operation op. It assumes that op(0,0)==0

source

SparseArrays.nnz — Method.

nnz(lnk)

Return number of nonzero entries.

source

ExtendableSparse.colptrs — Method.

colptrs(ext)

flush! and return colptr of in ext.cscmatrix.

source

ExtendableSparse.flush! — Method.

If there are new entries in extension, create new CSC matrix by adding the cscmatrix and linked list matrix and reset the linked list based extension.

source

ExtendableSparse.updateindex! — Method.

updateindex!(ext, op, v, i, j)

Update element of the matrix with operation op. This can replace the following code and save one index search during acces:

using ExtendableSparse # hide
A=ExtendableSparseMatrix(3,3)
A[1,2]+=0.1
A

using ExtendableSparse # hide

A=ExtendableSparseMatrix(3,3)
updateindex!(A,+,0.1,1,2)
A

source

SparseArrays.nnz — Method.

nnz(ext)

flush! and return number of nonzeros in ext.cscmatrix.

source

ExtendableSparse.fdrand! — Method.

fdrand!(A, nx, ny, nz; update, rand)

After setting all nonzero entries to zero, fill resp. update matrix with finite difference discretization data on a unit hypercube. See fdrand for documentation of the parameters.

It is required that size(A)==(N,N) where N=nx*ny*nz

source

ExtendableSparse.fdrand — Method.

fdrand(nx, ny, nz; matrixtype, update, rand)

Create matrix for a mock finite difference operator for a diffusion problem with random coefficients on a unit hypercube $\Omega\subset\mathbb R^d$. with $d=1$ if nx>1 && ny==1 && nz==1, $d=2$ if nx>1 && ny>1 && nz==1 and $d=3$ if nx>1 && ny>1 && nz>1 .

\[ \begin{align*} -\nabla a \nabla u &= f&& \text{in}\; \Omega \\ a\nabla u\cdot \vec n + bu &=g && \text{on}\; \partial\Omega \end{align*}\]

The matrix is irreducibly diagonally dominant, has positive main diagonal entries and nonpositive off-diagonal entries, hence it has the M-Property. Therefore, its inverse will be a dense matrix with positive entries. Moreover it is symmmetric, and positive definite.

Parameters+ default values:

Parameter + default vale	Description

nx	Number of unknowns in x direction

ny	Number of unknowns in y direction

nz	Number of unknowns in z direction

matrixtype = SparseMatrixCSC	Matrix type

update = (A,v,i,j)-> A[i,j]+=v	Element update function

rand =()-> rand()	Random number generator

The sparsity structure is fixed to an orthogonal grid, resulting in a 3, 5 or 7 diagonal matrix depending on dimension. The entries are random unless e.g. rand=()->1 is passed as random number generator. Tested for Matrix, SparseMatrixCSC and ExtendableSparseMatrix.

source

ExtendableSparse.sprand! — Method.

sprand!(A, xnnz)

Fill empty sparse matrix A with random nonzero elements from interval [1,2] using incremental assembly.

source

ExtendableSparse.sprand_sdd! — Method.

Fill sparse matrix with random entries such that it becomes strictly diagonally dominant and thus invertible and has a fixed number nnzrow (default: 4) of nonzeros in its rows. The matrix bandwidth is bounded by sqrt(n) in order to resemble a typical matrix of a 2D piecewise linear FEM discretization.

source

Base.:+ — Method.

+(lnk, csc)

Add SparseMatrixCSC matrix and SparseMatrixLNK lnk, returning a SparseMatrixCSC

source

Base.getindex — Method.

getindex(lnk, i, j)

Return value stored for entry or zero if not found

source

Base.setindex! — Method.

setindex!(lnk, _v, _i, _j)

Update value of existing entry, otherwise extend matrix if _v is nonzero.

source

Base.size — Method.

size(lnk)

Return tuple containing size of the matrix.

source

Base.:+ — Method.

Add SparseMatrixCSC matrix and ExtendableSparseMatrix ext.

source

Base.:- — Method.

-(ext, csc)

Subtract SparseMatrixCSC matrix from ExtendableSparseMatrix ext.

source

Base.:- — Method.

-(csc, ext)

Subtract ExtendableSparseMatrix ext from SparseMatrixCSC.

source

Base.:\ — Method.

\(ext, X)

\ for extmatrix

source

Base.getindex — Method.

getindex(ext, i, j)

Find index in CSC matrix and return value, if it exists. Otherwise, return value from extension.

source

Base.size — Method.

size(ext)

Size of ExtendableSparseMatrix.

source

LinearAlgebra.cond — Function.

cond(A)
cond(A, p)

flush! and calculate cond from cscmatrix

source

LinearAlgebra.ldiv! — Method.

ldiv!(r, ext, x)

flush! and ldiv with ext.cscmatrix

source

LinearAlgebra.ldiv! — Method.

ldiv!(r, ext, x)

flush! and ldiv with ext.cscmatrix

source

LinearAlgebra.lu — Method.

lu(ext)

flush! and return LU factorization of ext.cscmatrix

source

LinearAlgebra.mul! — Method.

mul!(r, ext, x)

flush! and multiply with ext.cscmatrix

source

LinearAlgebra.mul! — Method.

mul!(r, ext, x)

flush! and multiply with ext.cscmatrix

source

LinearAlgebra.norm — Function.

norm(A)
norm(A, p)

flush! and calculate norm from cscmatrix

source

LinearAlgebra.opnorm — Function.

opnorm(A)
opnorm(A, p)

flush! and calculate opnorm from cscmatrix

source

SparseArrays.findnz — Method.

findnz(ext)

flush! and return findnz(ext.cscmatrix).

source

SparseArrays.nonzeros — Method.

nonzeros(ext)

flush! and return nonzeros in ext.cscmatrix.

source

SparseArrays.rowvals — Method.

rowvals(ext)

flush! and return rowvals in ext.cscmatrix.

source