Sparse matrix handling

Matrix creation and update API

ExtendableSparse.ExtendableSparseMatrix — Type

mutable struct ExtendableSparseMatrix{Tv, Ti<:Integer} <: SparseArrays.AbstractSparseArray{Tv, Ti<:Integer, 2}

Extendable sparse matrix. A nonzero entry of this matrix is contained either in cscmatrix, or in lnkmatrix, never in both.

cscmatrix::SparseArrays.SparseMatrixCSC: Final matrix data

lnkmatrix::Union{Nothing, SparseMatrixLNK{Tv, Ti}} where {Tv, Ti<:Integer}: Linked list structure holding data of extension

phash::UInt64: Pattern hash

ExtendableSparse.ExtendableSparseMatrix — Method

ExtendableSparseMatrix(A)

Create ExtendableSparseMatrix from AbstractMatrix, dropping all zero entries. This is the equivalent to sparse(A).

ExtendableSparse.ExtendableSparseMatrix — Method

ExtendableSparseMatrix(I,J,V)
ExtendableSparseMatrix(I,J,V,m,n)
ExtendableSparseMatrix(I,J,V,combine::Function)
ExtendableSparseMatrix(I,J,V,m,n,combine::Function)

Create ExtendableSparseMatrix from triplet (COO) data.

ExtendableSparse.ExtendableSparseMatrix — Method

ExtendableSparseMatrix(D)

Create ExtendableSparseMatrix from Diagonal

ExtendableSparse.ExtendableSparseMatrix — Method

ExtendableSparseMatrix(csc)

Create ExtendableSparseMatrix from SparseMatrixCSC

ExtendableSparse.ExtendableSparseMatrix — Method

ExtendableSparseMatrix(Tv,Ti,m,n)
ExtendableSparseMatrix(Tv,m,n)
ExtendableSparseMatrix(m,n)

Create empty ExtendableSparseMatrix. This is equivalent to spzeros(m,n) for SparseMartrixCSC.

Base.:* — Method

*(ext, d)

Base.:* — Method

*(d, ext)

Base.:+ — Method

+(ext, csc)

Add SparseMatrixCSC matrix and ExtendableSparseMatrix ext.

Base.:- — Method

-(ext, csc)

Subtract SparseMatrixCSC matrix from ExtendableSparseMatrix ext.

Base.:- — Method

-(csc, ext)

Subtract ExtendableSparseMatrix ext from SparseMatrixCSC.

Base.:\ — Method

\(symm_ext, B)

\ for Hermitian{ExtendableSparse}

Base.:\ — Method

\(symm_ext, B)

\ for Symmetric{ExtendableSparse}

Base.:\ — Method

A

\ for ExtendableSparse. It calls the LU factorization form Sparspak.jl, unless GPL components are allowed in the Julia sysimage and the floating point type of the matrix is Float64 or Complex64. In that case, Julias standard `` is called, which is realized via UMFPACK.

Base.copy — Method

copy(S)

Base.getindex — Method

getindex(ext, i, j)

Find index in CSC matrix and return value, if it exists. Otherwise, return value from extension.

Base.setindex! — Method

setindex!(ext, v, i, j)

Find index in CSC matrix and set value if it exists. Otherwise, set index in extension if v is nonzero.

Base.show — Method

show(io, _, ext)

Show matrix

Base.similar — Method

similar(m)

Create similar but emtpy extendableSparseMatrix

Base.size — Method

size(ext)

Size of ExtendableSparseMatrix.

ExtendableSparse.flush! — Method

flush!(ext)

If there are new entries in extension, create new CSC matrix by adding the cscmatrix and linked list matrix and reset the linked list based extension.

ExtendableSparse.rawupdateindex! — Method

rawupdateindex!(ext, op, v, i, j)

Like updateindex! but without checking if v is zero.

ExtendableSparse.updateindex! — Method

updateindex!(ext, op, v, i, j)

Update element of the matrix with operation op. This can replace the following code and save one index search during acces:

using ExtendableSparse # hide
A=ExtendableSparseMatrix(3,3)
A[1,2]+=0.1
A

using ExtendableSparse # hide

A=ExtendableSparseMatrix(3,3)
updateindex!(A,+,0.1,1,2)
A

If v is zero, no new entry is created.

LinearAlgebra.cond — Function

cond(A)
cond(A, p)

flush! and calculate cond from cscmatrix

LinearAlgebra.issymmetric — Method

issymmetric(A)

flush! and check for symmetry of cscmatrix

LinearAlgebra.ldiv! — Method

ldiv!(r, ext, x)

flush! and ldiv with ext.cscmatrix

LinearAlgebra.mul! — Method

mul!(r, ext, x)

flush! and multiply with ext.cscmatrix

LinearAlgebra.norm — Function

norm(A)
norm(A, p)

flush! and calculate norm from cscmatrix

LinearAlgebra.opnorm — Function

opnorm(A)
opnorm(A, p)

flush! and calculate opnorm from cscmatrix

SparseArrays.dropzeros! — Method

dropzeros!(ext)

SparseArrays.findnz — Method

findnz(ext)

flush! and return findnz(ext.cscmatrix).

SparseArrays.getcolptr — Method

getcolptr(ext)

flush! and return colptr of in ext.cscmatrix.

SparseArrays.nnz — Method

nnz(ext)

flush! and return number of nonzeros in ext.cscmatrix.

SparseArrays.nonzeros — Method

nonzeros(ext)

flush! and return nonzeros in ext.cscmatrix.

SparseArrays.rowvals — Method

rowvals(ext)

flush! and return rowvals in ext.cscmatrix.

SparseArrays.sparse — Method

sparse(A)

Create SparseMatrixCSC from ExtendableSparseMatrix

Test matrix creation

ExtendableSparse.fdrand! — Method

fdrand!(A, nx)
fdrand!(A, nx, ny)
fdrand!(A, nx, ny, nz; update, rand)

After setting all nonzero entries to zero, fill resp. update matrix with finite difference discretization data on a unit hypercube. See fdrand for documentation of the parameters.

It is required that size(A)==(N,N) where N=nx*ny*nz

ExtendableSparse.fdrand — Method

fdrand(, nx)
fdrand(, nx, ny)
fdrand(, nx, ny, nz; matrixtype, update, rand, symmetric)
fdrand(nx)

Create matrix for a mock finite difference operator for a diffusion problem with random coefficients on a unit hypercube $\Omega\subset\mathbb R^d$. with $d=1$ if nx>1 && ny==1 && nz==1, $d=2$ if nx>1 && ny>1 && nz==1 and $d=3$ if nx>1 && ny>1 && nz>1 . In the symmetric case it corresponds to

\[ \begin{align*} -\nabla a \nabla u &= f&& \text{in}\; \Omega \\ a\nabla u\cdot \vec n + bu &=g && \text{on}\; \partial\Omega \end{align*}\]

The matrix is irreducibly diagonally dominant, has positive main diagonal entries and nonpositive off-diagonal entries, hence it has the M-Property. Therefore, its inverse will be a dense matrix with positive entries, and the spectral radius of the Jacobi iteration matrix $ho(I-D(A)^{-1}A)<1$ .

Moreover, in the symmetric case, it is positive definite.

Parameters+ default values:

Parameter + default vale	Description

nx	Number of unknowns in x direction

ny	Number of unknowns in y direction

nz	Number of unknowns in z direction

matrixtype = SparseMatrixCSC	Matrix type

update = (A,v,i,j)-> A[i,j]+=v	Element update function

rand =()-> rand()	Random number generator

symmetric=true	Whether to create symmetric matrix or not

The sparsity structure is fixed to an orthogonal grid, resulting in a 3, 5 or 7 diagonal matrix depending on dimension. The entries are random unless e.g. rand=()->1 is passed as random number generator. Tested for Matrix, SparseMatrixCSC, ExtendableSparseMatrix, Tridiagonal, SparseMatrixLNK and :COO

ExtendableSparse.sprand! — Method

sprand!(A, xnnz)

Fill empty sparse matrix A with random nonzero elements from interval [1,2] using incremental assembly.

ExtendableSparse.sprand_sdd! — Method

sprand_sdd!(A; nnzrow)

Fill sparse matrix with random entries such that it becomes strictly diagonally dominant and thus invertible and has a fixed number nnzrow (default: 4) of nonzeros in its rows. The matrix bandwidth is bounded by sqrt(n) in order to resemble a typical matrix of a 2D piecewise linear FEM discretization.