Factorizations & Preconditioners

Factorizations

In this package, preconditioners and LU factorizations are subcategories are both seen as complete or approximate factorizations. Correspondingly there is a common API for their creation.

Factorizations from these package know the matrices which have been factorized.

ExtendableSparse.default_options — Constant

const default_options

Default options for various factorizations:

kind: umfpack

ensurelu: false

droptol: 0.001

ExtendableSparse.AbstractExtendableSparseFactorization — Type

abstract type AbstractExtendableSparseFactorization{Tv, Ti}

Abstract type for a factorization with ExtandableSparseMatrix.

Any such preconditioner should have the following fields

  A
  fact
  phash

and methods

  factorize(A; kwargs)
  ldiv!(u, precon,v)
  ldiv!(precon,v)
  issolver(precon)
  update!(precon)
  factorize!(precon,A; kwargs)

The idea is that, depending if the matrix pattern has changed, different steps are needed to update the preconditioner.

Moreover, they have the ExtendableSparseMatrix as a field, ensuring consistency after construction.

ExtendableSparse.AbstractExtendableSparseLU — Type

abstract type AbstractExtendableSparseLU{Tv, Ti} <: ExtendableSparse.AbstractExtendableSparseFactorization{Tv, Ti}

Abstract subtype for (full) LU factorizations

ExtendableSparse.AbstractExtendableSparsePreconditioner — Type

abstract type AbstractExtendableSparsePreconditioner{Tv, Ti} <: ExtendableSparse.AbstractExtendableSparseFactorization{Tv, Ti}

Abstract subtype for preconditioners

Base.:\ — Method

 lufacths

Solve LU factorization problem.

ExtendableSparse.factorize! — Method

factorize!(factorization_or_nothing, matrix; kwargs...)

Update factorization, possibly reusing information from the current state. This method is aware of pattern changes.

If nothing is passed as first parameter, factorize is called.

ExtendableSparse.factorize — Method

factorize(matrix)
factorize(matrix; kind=:umfpack)

Default Julia LU factorization based on UMFPACK.

factorize(matrix; kind=:pardiso)
factorize(matrix; kind=:mklpardiso)

LU factorization based on pardiso. For using this, you need to issue using Pardiso. :pardiso uses the solver from pardiso-project.org, while :mklpardiso uses the early 2000's fork in Intel's MKL library.

factorize(matrix; kind=:ilu0)

Create the ILU0Preconditioner from this package.

factorize(matrix; kind=:jacobi)

Create the JacobiPreconditioner from this package.

factorize(matrix; kind=:pjacobi)

Create the ParallelJacobiPreconditioner from this package.

factorize(matrix; kind=:ilut, droptol=1.0e-3)

Create the ILUTPreconditioner wrapping the one from IncompleteLU.jl For using this, you need to issue using IncompleteLU

factorize(matrix; kind=:rsamg)

Create the AMGPreconditioner wrapping the Ruge-Stüben AMG preconditioner from AlgebraicMultigrid.jl

ExtendableSparse.issolver — Method

issolver(factorization)

Determine if factorization is a solver or not

ExtendableSparse.options — Method

options(;kwargs...)

Set default options and blend them with kwargs

ExtendableSparse.update! — Method

update!(factorization)

Update factorization after matrix update.

LinearAlgebra.ldiv! — Method

ldiv!(u,factorization,v)
ldiv!(factorization,v)

Solve factorization.

LinearAlgebra.lu! — Method

lu!(factorization_or_nothing, matrix; kwargs...)

Update LU factorization, possibly reusing information from the current state. This method is aware of pattern changes.

If nothing is passed as first parameter, factorize is called.

LinearAlgebra.lu — Method

lu(matrix)
lu(matrix,kind=:pardiso)
lu(matrix,kind=:mklpardiso)

Wrapper for factorize restricted to lu factorizations.

LU Factorizations

ExtendableSparse.ExtendableSparseUmfpackLU — Type

mutable struct ExtendableSparseUmfpackLU{Tv, Ti} <: ExtendableSparse.AbstractExtendableSparseLU{Tv, Ti}

Default Julia LU Factorization based on umfpack.

ExtendableSparse.ExtendableSparseUmfpackLU — Method

ExtendableSparseUmfpackLU(A)

ExtendableSparse.PardisoLU — Type

mutable struct PardisoLU{Tv, Ti} <: ExtendableSparse.AbstractExtendableSparseLU{Tv, Ti}

LU Factorization

ExtendableSparse.PardisoLU — Method

PardisoLU(A; ps=Pardiso.MKLPardisoSolver)

Preconditioners

ExtendableSparse.JacobiPreconditioner — Type

mutable struct JacobiPreconditioner{Tv, Ti} <: ExtendableSparse.AbstractExtendableSparsePreconditioner{Tv, Ti}

Jacobi preconditoner

ExtendableSparse.JacobiPreconditioner — Method

JacobiPreconditioner(A)
JacobiPreconditioner(cscmatrix)

ExtendableSparse.ParallelJacobiPreconditioner — Type

mutable struct ParallelJacobiPreconditioner{Tv, Ti} <: ExtendableSparse.AbstractExtendableSparsePreconditioner{Tv, Ti}

Parallel Jacobi preconditioner

ExtendableSparse.ParallelJacobiPreconditioner — Method

ParallelJacobiPreconditioner(A)
ParallelJacobiPreconditioner(cscmatrix)

ExtendableSparse.ILU0Preconditioner — Type

mutable struct ILU0Preconditioner{Tv, Ti} <: ExtendableSparse.AbstractExtendableSparsePreconditioner{Tv, Ti}

ILU(0) Preconditioner

ExtendableSparse.ILU0Preconditioner — Method

ILU0Preconditioner(extsparse)
ILU0Preconditioner(cscmatrix)

ExtendableSparse.ILUTPreconditioner — Type

mutable struct ILUTPreconditioner{Tv, Ti} <: ExtendableSparse.AbstractExtendableSparsePreconditioner{Tv, Ti}

ILU(T) preconditioner

ExtendableSparse.ILUTPreconditioner — Method

ILUTPreconditioner(matrix; droptol=1.0e-3)

ExtendableSparse.AMGPreconditioner — Type

mutable struct AMGPreconditioner{Tv, Ti} <: ExtendableSparse.AbstractExtendableSparsePreconditioner{Tv, Ti}

ExtendableSparse.AMGPreconditioner — Method

AMGPreconditioner(extmatrix)