Factorizations & Preconditioners

This functionality probably will be reduced in favor of LinearSolve.jl.

Factorizations

In this package, preconditioners and LU factorizations are both seen as complete or approximate factorizations. Correspondingly we provide a common API for them.

ExtendableSparse.factorize! — Method

factorize!(factorization, matrix)

Update or create factorization, possibly reusing information from the current state. This method is aware of pattern changes.

ExtendableSparse.issolver — Method

issolver(factorization)

Determine if factorization is a solver or not

ExtendableSparse.update! — Method

update!(factorization)

Update factorization after matrix update.

ExtendableSparse.AbstractFactorization — Type

abstract type AbstractFactorization

Abstract type for a factorization with ExtandableSparseMatrix.

This type is meant to be a "type flexible" (with respect to the matrix element type) and lazily construcdet (can be constructed without knowing the matrix, and updated later) LU factorization or preconditioner. It wraps different concrete, type fixed factorizations which shall provide the usual ldiv! methods.

Any such preconditioner/factorization MyFact should have the following fields

  A::ExtendableSparseMatrix
  factorization
  phash::UInt64

and provide methods

  MyFact(;kwargs...) 
  update!(precon::MyFact)

The idea is that, depending if the matrix pattern has changed, different steps are needed to update the preconditioner.

LU Factorizations

Handling of the LU factorizations is meant to support a workflow where sequences of problems are solved based on the same matrix, where one possibly wants to re-use existing symbolic factorization data.

The support comes in two flavors.

Using factorize! which can work as a drop-in replacement for lu!:

using ExtendableSparse, LinearAlgebra
using Pardiso
A = fdrand(20, 20, 1; matrixtype = ExtendableSparseMatrix)
n = size(A, 1)
b = rand(n)
factorization = MKLPardisoLU()
factorize!(factorization, A)
nm1 = norm(factorization \ b)

# mock update from Newton etc.
for i = 4:(n - 3)
    A[i, i + 3] -= 1.0e-4
end
factorize!(factorization, A)
nm2 = norm(factorization \ b)
nm1, nm2

(1907.2705094925202, 1943.6798954198168)

Using update!, where the matrix only needs to be given at construction time:

using ExtendableSparse, LinearAlgebra
A = fdrand(20, 20, 1; matrixtype = ExtendableSparseMatrix)
n = size(A, 1)
b = rand(n)
factorization = CholeskyFactorization(A)
nm1 = norm(factorization \ b)

# mock update from Newton etc.
for i = 4:(n - 3)
    A[i, i + 3] -= 1.0e-4
    A[i - 3, i] -= 1.0e-4
end
update!(factorization)
nm2 = norm(factorization \ b)
nm1, nm2

(1926.1291936005678, 2080.0353048864035)

API

Available by default:

ExtendableSparse.LUFactorization — Type

LUFactorization()
LUFactorization(matrix)

Default LU Factorization. Maps to Sparspak.jl for non-GPL builds, otherwise to UMFPACK.

ExtendableSparse.SparspakLU — Type

SparspakLU() 
SparspakLU(matrix) 

LU factorization based on Sparspak.jl (P.Krysl's Julia re-implementation of Sparspak by George & Liu)

Pardiso extension:

ExtendableSparse.PardisoLU — Function

PardisoLU(;iparm::Vector, 
           dparm::Vector, 
           mtype::Int)

PardisoLU(matrix; iparm,dparm,mtype)

LU factorization based on pardiso. For using this, you need to issue using Pardiso and have the pardiso library from pardiso-project.org installed.

The optional keyword arguments mtype, iparm and dparm are Pardiso internal parameters.

Forsetting them, one can also access the PardisoSolver e.g. like

using Pardiso
plu=PardisoLU()
Pardiso.set_iparm!(plu.ps,5,13.0)

ExtendableSparse.MKLPardisoLU — Function

MKLPardisoLU(;iparm::Vector, mtype::Int)

MKLPardisoLU(matrix; iparm, mtype)

LU factorization based on pardiso. For using this, you need to issue using Pardiso. This version uses the early 2000's fork in Intel's MKL library.

The optional keyword arguments mtype and iparm are Pardiso internal parameters.

For setting them you can also access the PardisoSolver e.g. like

using Pardiso
plu=MKLPardisoLU()
Pardiso.set_iparm!(plu.ps,5,13.0)

Preconditioners

The API is similar to that for LU factorizations.

The support comes in two flavors.

Using factorize!:

using ExtendableSparse, LinearAlgebra
using IterativeSolvers, IncompleteLU
A = fdrand(20, 20, 1; matrixtype = ExtendableSparseMatrix)
n = size(A, 1)
b = rand(n)
preconditioner = ILUTPreconditioner(; droptol = 1.0e-2)
factorize!(preconditioner, A)

# mock update from Newton etc.
nm1 = norm(bicgstabl(A, b, 1; Pl = preconditioner))
for i = 4:(n - 3)
    A[i, i + 3] -= 1.0e-4
end
factorize!(preconditioner, A)
nm2 = norm(bicgstabl(A, b, 1; Pl = preconditioner))
nm1, nm2

(1769.227910232084, 1802.182610575726)

Using update!:

using ExtendableSparse, LinearAlgebra
using IterativeSolvers
A = fdrand(20, 20, 1; matrixtype = ExtendableSparseMatrix)
n = size(A, 1)
b = rand(n)
preconditioner = ILU0Preconditioner(A)
nm1 = norm(cg(A, b; Pl = preconditioner))

# mock update from Newton etc.
for i = 4:(n - 3)
    A[i, i + 3] -= 1.0e-4
    A[i - 3, i] -= 1.0e-4
end
update!(preconditioner)
nm2 = norm(cg(A, b; Pl = preconditioner))
nm1, nm2

(1924.4903739351203, 2003.8579299688224)

API

Recommended

Available by default:

ExtendableSparse.ILUZeroPreconditioner — Type

ILUZeroPreconditioner()
ILUZeroPreconditioner(matrix)

Incomplete LU preconditioner with zero fill-in using ILUZero.jl. This preconditioner also calculates and stores updates to the off-diagonal entries and thus delivers better convergence than the ILU0Preconditioner.

ExtendableSparse.PointBlockILUZeroPreconditioner — Type

PointBlockILUZeroPreconditioner(;blocksize)
PointBlockILUZeroPreconditioner(matrix;blocksize)

Incomplete LU preconditioner with zero fill-in using ILUZero.jl. This preconditioner also calculates and stores updates to the off-diagonal entries and thus delivers better convergence than the ILU0Preconditioner.

ExtendableSparse.BlockPreconditioner — Type

 BlockPreconditioner(;partitioning, factorization=LUFactorization)

Create a block preconditioner from partition of unknowns given by partitioning, a vector of AbstractVectors describing the indices of the partitions of the matrix. For a matrix of size n x n, e.g. partitioning could be [ 1:n÷2, (n÷2+1):n] or [ 1:2:n, 2:2:n]. Factorization is a callable (Function or struct) which allows to create a factorization (with ldiv! methods) from a submatrix of A.

ExtendableSparse.allow_views — Method

allow_views(::preconditioner_type)

Factorizations on matrix partitions within a block preconditioner may or may not work with array views. E.g. the umfpack factorization cannot work with views, while ILUZeroPreconditioner can. Implementing a method for allow_views returning false resp. true allows to dispatch to the proper case.

Extensions:

ExtendableSparse.ILUTPreconditioner — Function

ILUTPreconditioner(;droptol=1.0e-3)
ILUTPreconditioner(matrix; droptol=1.0e-3)

Create the ILUTPreconditioner wrapping the one from IncompleteLU.jl For using this, you need to issue using IncompleteLU.

ExtendableSparse.AMGPreconditioner — Function

AMGPreconditioner(;max_levels=10, max_coarse=10)
AMGPreconditioner(matrix;max_levels=10, max_coarse=10)

Create the AMGPreconditioner wrapping the Ruge-Stüben AMG preconditioner from AlgebraicMultigrid.jl

Warning

Deprecated in favor of RS_AMGPreconditioner

Experimental

ExtendableSparse.JacobiPreconditioner — Type

JacobiPreconditioner()
JacobiPreconditioner(matrix)

Jacobi preconditioner.

ExtendableSparse.ParallelJacobiPreconditioner — Type

ParallelJacobiPreconditioner()
ParallelJacobiPreconditioner(matrix)

ParallelJacobi preconditioner.

ExtendableSparse.ILU0Preconditioner — Type

ILU0Preconditioner()
ILU0Preconditioner(matrix)

Incomplete LU preconditioner with zero fill-in, without modification of off-diagonal entries, so it delivers slower convergende than ILUZeroPreconditioner.

ExtendableSparse.ParallelILU0Preconditioner — Type

ParallelILU0Preconditioner()
ParallelILU0Preconditioner(matrix)

Parallel ILU preconditioner with zero fill-in.