Factorizations & Preconditioners

This functionality probably will be reduced in favor of LinearSolve.jl.

Factorizations

In this package, preconditioners and LU factorizations are both seen as complete or approximate factorizations. Correspondingly we provide a common API for their creation.

ExtendableSparse.factorize! — Method

factorize!(factorization, matrix)

Update or create factorization, possibly reusing information from the current state. This method is aware of pattern changes.

ExtendableSparse.issolver — Method

issolver(factorization)

Determine if factorization is a solver or not

ExtendableSparse.update! — Method

update!(factorization)

Update factorization after matrix update.

ExtendableSparse.AbstractFactorization — Type

abstract type AbstractFactorization{Tv, Ti}

Abstract type for a factorization with ExtandableSparseMatrix.

Any such preconditioner should have the following fields

  A::ExtendableSparseMatrix
  fact
  phash::UInt64

and provide a method

  update!(precon)

The idea is that, depending if the matrix pattern has changed, different steps are needed to update the preconditioner.

Moreover, they have the ExtendableSparseMatrix A as a field and an update! method, ensuring consistency after construction.

LU Factorizations

Handling of the LU factorizations is meant to support a workflow where sequences of problems are solved based on the same matrix, where one possibly wants to re-use existing symbolic factorization data.

The support comes in two flavors.

Using factorize! which can work as a drop-in replacement for lu!:

using ExtendableSparse, LinearAlgebra
using Pardiso
A = fdrand(20, 20, 1; matrixtype = ExtendableSparseMatrix)
n = size(A, 1)
b = rand(n)
factorization = MKLPardisoLU()
factorize!(factorization, A)
nm1 = norm(factorization \ b)

# mock update from Newton etc.
for i = 4:(n - 3)
    A[i, i + 3] -= 1.0e-4
end
factorize!(factorization, A)
nm2 = norm(factorization \ b)
nm1, nm2

(1995.6588863934142, 2038.996840155146)

Using update!, where the matrix only needs to be given at construction time:

using ExtendableSparse, LinearAlgebra
A = fdrand(20, 20, 1; matrixtype = ExtendableSparseMatrix)
n = size(A, 1)
b = rand(n)
factorization = CholeskyFactorization(A)
nm1 = norm(factorization \ b)

# mock update from Newton etc.
for i = 4:(n - 3)
    A[i, i + 3] -= 1.0e-4
    A[i - 3, i] -= 1.0e-4
end
update!(factorization)
nm2 = norm(factorization \ b)
nm1, nm2

(2104.5797347183675, 2300.3170738473927)

API

ExtendableSparse.LUFactorization — Method

LUFactorization(;valuetype=Float64, indextype=Int64)
LUFactorization(matrix)

Default Julia LU Factorization based on umfpack.

ExtendableSparse.MKLPardisoLU — Method

MKLPardisoLU(;valuetype=Float64, 
           indextype=Int64,
           iparm::Vector, 
           mtype::Int)

MKLPardisoLU(matrix; iparm, mtype)

LU factorization based on pardiso. For using this, you need to issue using Pardiso. This version uses the early 2000's fork in Intel's MKL library.

The optional keyword arguments mtype and iparm are (Pardiso internal parameters)[https://github.com/JuliaSparse/Pardiso.jl#readme].

For setting them you can also access the PardisoSolver e.g. like

using Pardiso
plu=MKLPardisoLU()
Pardiso.set_iparm!(plu.ps,5,13.0)

ExtendableSparse.PardisoLU — Method

PardisoLU(;valuetype=Float64, 
           indextype=Int64,
           iparm::Vector, 
           dparm::Vector, 
           mtype::Int)

PardisoLU(matrix; iparm,dparm,mtype)

LU factorization based on pardiso. For using this, you need to issue using Pardiso and have the pardiso library from pardiso-project.org installed.

The optional keyword arguments mtype, iparm and dparm are (Pardiso internal parameters)[https://github.com/JuliaSparse/Pardiso.jl#readme].

Forsetting them, one can also access the PardisoSolver e.g. like

using Pardiso
plu=PardisoLU()
Pardiso.set_iparm!(plu.ps,5,13.0)

ExtendableSparse.SparspakLU — Method

SparspakLU(;valuetype=Float64) 

LU factorization based on Sparspak.jl (P.Krysl's Julia re-implementation of Sparspak by George & Liu)

Preconditioners

The API is similar to that for LU factorizations.

The support comes in two flavors.

Using factorize!:

using ExtendableSparse, LinearAlgebra
using IterativeSolvers, IncompleteLU
A = fdrand(20, 20, 1; matrixtype = ExtendableSparseMatrix)
n = size(A, 1)
b = rand(n)
preconditioner = ILUTPreconditioner(; droptol = 1.0e-2)
factorize!(preconditioner, A)

# mock update from Newton etc.
nm1 = norm(bicgstabl(A, b, 1; Pl = preconditioner))
for i = 4:(n - 3)
    A[i, i + 3] -= 1.0e-4
end
factorize!(preconditioner, A)
nm2 = norm(bicgstabl(A, b, 1; Pl = preconditioner))
nm1, nm2

(1993.6185750392685, 2033.5695175328874)

Using update!:

using ExtendableSparse, LinearAlgebra
using IterativeSolvers
A = fdrand(20, 20, 1; matrixtype = ExtendableSparseMatrix)
n = size(A, 1)
b = rand(n)
preconditioner = ILU0Preconditioner(A)
nm1 = norm(cg(A, b; Pl = preconditioner))

# mock update from Newton etc.
for i = 4:(n - 3)
    A[i, i + 3] -= 1.0e-4
    A[i - 3, i] -= 1.0e-4
end
update!(preconditioner)
nm2 = norm(cg(A, b; Pl = preconditioner))
nm1, nm2

(1914.3464710812802, 1990.641124381375)

API

ExtendableSparse.JacobiPreconditioner — Method

JacobiPreconditioner(;valuetype=Float64, indextype=Int64)
JacobiPreconditioner(matrix)

Jacobi preconditioner

ExtendableSparse.ParallelJacobiPreconditioner — Method

ParallelJacobiPreconditioner(;valuetype=Float64, indextype=Int64)
ParallelJacobiPreconditioner(matrix)

Parallel Jacobi preconditioner

ExtendableSparse.ILUZeroPreconditioner — Method

ILUZeroPreconditioner(;valuetype=Float64,indextype=Int64)
ILUZeroPreconditioner(matrix)

Incomplete LU preconditioner with zero fill-in using ILUZero.jl. This preconditioner also calculates and stores updates to the off-diagonal entries and thus delivers better convergence than the ILU0Preconditoner.

ExtendableSparse.ILU0Preconditioner — Method

ILU0Preconditioner(;valuetype=Float64,indextype=Int64)
ILU0Preconditioner(matrix)

Incomplete LU preconditioner with zero fill-in, without modification of off-diagonal entries, so it delivers slower convergende than ILUZeroPreconditoner.

ExtendableSparse.ParallelILU0Preconditioner — Method

ParallelILU0Preconditioner(;valuetype=Float64,indextype=Int64)
ParallelILU0Preconditioner(matrix)

Parallel ILU preconditioner with zero fill-in.

ExtendableSparse.ILUTPreconditioner — Method

ILUTPreconditioner(;droptol=1.0e-3,valuetype=Float64, indextype=Int64)
ILUTPreconditioner(matrix; droptol=1.0e-3)

Create the ILUTPreconditioner wrapping the one from IncompleteLU.jl For using this, you need to issue using IncompleteLU.

ExtendableSparse.AMGPreconditioner — Method

AMGPreconditioner(;max_levels=10, max_coarse=10, valuetype=Float64,indextype=Int64)
AMGPreconditioner(matrix;max_levels=10, max_coarse=10)

Create the AMGPreconditioner wrapping the Ruge-Stüben AMG preconditioner from AlgebraicMultigrid.jl