Factorizations & Preconditioners

Factorizations

In this package, preconditioners and LU factorizations are both seen as complete or approximate factorizations. Correspondingly we provide a common API for their creation.

ExtendableSparse.factorize! — Method

factorize!(factorization, matrix)

Update or create factorization, possibly reusing information from the current state. This method is aware of pattern changes.

ExtendableSparse.issolver — Method

issolver(factorization)

Determine if factorization is a solver or not

ExtendableSparse.update! — Method

update!(factorization)

Update factorization after matrix update.

ExtendableSparse.AbstractFactorization — Type

abstract type AbstractFactorization{Tv, Ti}

Abstract type for a factorization with ExtandableSparseMatrix.

Any such preconditioner should have the following fields

  A::ExtendableSparseMatrix
  fact
  phash::UInt64

and provide a method

  update!(precon)

The idea is that, depending if the matrix pattern has changed, different steps are needed to update the preconditioner.

Moreover, they have the ExtendableSparseMatrix as a field, ensuring consistency after construction.

LU Factorizations

Handling of the LU factorizations is meant to support a workflow where sequences of problems are solved based on the same matrix, where one possibly wants to re-use existing symbolic factorization data.

The support comes in two flavors.

Using factorize! which can work as a drop-in replacement for lu!:

using ExtendableSparse, LinearAlgebra
using Pardiso
A=fdrand(20,20,1,matrixtype=ExtendableSparseMatrix)
n=size(A,1)
b=rand(n)
factorization=MKLPardisoLU()
factorize!(factorization,A)
nm1=norm(factorization\b)

# mock update from Newton etc.
for i=4:n-3
    A[i,i+3]-=1.0e-4
end
factorize!(factorization,A)
nm2=norm(factorization\b)
nm1,nm2

(2130.6818702910755, 2176.289253388951)

Using update!, where the matrix only needs to be given at construction time:

using ExtendableSparse, LinearAlgebra
A=fdrand(20,20,1,matrixtype=ExtendableSparseMatrix)
n=size(A,1)
b=rand(n)
factorization=CholeskyFactorization(A)
nm1=norm(factorization\b)

# mock update from Newton etc.
for i=4:n-3
    A[i,i+3]-=1.0e-4
    A[i-3,i]-=1.0e-4
end
update!(factorization)
nm2=norm(factorization\b)
nm1,nm2

(1994.7488546110462, 2165.4166763142703)

API

ExtendableSparse.LUFactorization — Method

LUFactorization()
LUFactorization(matrix)

Default Julia LU Factorization based on umfpack.

ExtendableSparse.MKLPardisoLU — Method

MKLPardisoLU()
MKLPardisoLU(matrix)

LU factorization based on pardiso. For using this, you need to issue using Pardiso. This version uses the early 2000's fork in Intel's MKL library.

ExtendableSparse.PardisoLU — Method

LU factorization based on pardiso. For using this, you need to issue using Pardiso and have the pardiso library from pardiso-project.org installed.

Preconditioners

The API is similar to that for LU factorizations.

The support comes in two flavors.

Using factorize!:

using ExtendableSparse, LinearAlgebra
using IterativeSolvers,IncompleteLU
A=fdrand(20,20,1,matrixtype=ExtendableSparseMatrix)
n=size(A,1)
b=rand(n)
preconditioner=ILUTPreconditioner(droptol=1.0e-2)
factorize!(preconditioner,A)

# mock update from Newton etc.
nm1=norm(bicgstabl(A,b,1,Pl=preconditioner))
for i=4:n-3
    A[i,i+3]-=1.0e-4
end
factorize!(preconditioner,A)
nm2=norm(bicgstabl(A,b,1,Pl=preconditioner))
nm1,nm2

(2056.519839218806, 2099.079043953223)

Using update!:

using ExtendableSparse, LinearAlgebra
using IterativeSolvers
A=fdrand(20,20,1,matrixtype=ExtendableSparseMatrix)
n=size(A,1)
b=rand(n)
preconditioner=ILU0Preconditioner(A)
nm1=norm(cg(A,b,Pl=preconditioner))

# mock update from Newton etc.
for i=4:n-3
    A[i,i+3]-=1.0e-4
    A[i-3,i]-=1.0e-4
end
update!(preconditioner)
nm2=norm(cg(A,b,Pl=preconditioner))
nm1,nm2

(1823.8236950297057, 1893.0529637495206)

API

ExtendableSparse.JacobiPreconditioner — Method

JacobiPreconditioner()
JacobiPreconditioner(matrix)

Jacobi preconditioner

ExtendableSparse.ParallelJacobiPreconditioner — Method

ParallelJacobiPreconditioner()
ParallelJacobiPreconditioner(matrix)

Parallel Jacobi preconditioner

ExtendableSparse.ILU0Preconditioner — Method

ILU0Preconditioner()
ILU0Preconditioner(matrix)

ILU preconditioner with zero fill-in.

ExtendableSparse.ILUTPreconditioner — Method

ILUTPreconditioner(;droptol=1.0e-3)
ILUTPreconditioner(matrix; droptol=1.0e-3)

Create the ILUTPreconditioner wrapping the one from IncompleteLU.jl For using this, you need to issue using IncompleteLU.

ExtendableSparse.AMGPreconditioner — Method

AMGPreconditioner(;max_levels=10, max_coarse=10)
AMGPreconditioner(matrix;max_levels=10, max_coarse=10)

Create the AMGPreconditioner wrapping the Ruge-Stüben AMG preconditioner from AlgebraicMultigrid.jl